% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2BConvert.R
\name{B2BConvert}
\alias{B2BConvert}
\title{B2BConvert Converts a function of SNR into one of SNR, B2B, and Offset.}
\usage{
B2BConvert(f)
}
\arguments{
\item{f}{A function of a single argument \code{f( s)}. \code{f} can be a symbol, a
string, or an anonymous function.  If a symbol, the symbol name will
be remembered.  If a string, it will be passed to \code{match.fun} which
will return the function form that is the value of the named symbol
(i.e., not the symbol the string named).}
}
\value{
A function of three arguments \code{function( x, B2B, offset){...}}
where \code{x} is the symbol SNR, \code{B2B} is the back-to-back SNR (i.e. the
equivalent SNR when the input SNR (\code{x}) is infinite), and \code{offset} is
the offset to the function \code{f} (i.e., if \code{B2B} where infinite).
}
\description{
Creates a function \code{f( -dB( undB( -s) + undB( -B2B)) - offset)}
}
\details{
Note that all quantities are assumed to be in Decibels.
}
\examples{
QPSKdB.B2B <- B2BConvert( QPSKdB)

}
