% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2BQ.R
\docType{package}
\name{B2BQ-package}
\alias{B2BQ}
\alias{B2BQ-package}
\title{Package \code{B2BQ} will estimate an offset and B2B "Q" of a communications
system performance.}
\description{
High performance communications systems typically have inherent noise sources and other perfprmance limitations that need to be estimated. Measurments made at high signal to noise ratios typicaly result in zero errors due to limitation in available measurement time. The tools of this package help a user estimate the performance parameters. Includes theoretical performance functions for popular modulation methods, and utility functions to work with the methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PhilShea/B2BQ}
  \item Report bugs at \url{https://github.com/PhilShea/B2BQ/issues}
}

}
\keyword{internal}
