% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{BERDFc}
\alias{BERDFc}
\title{An example \code{BERDF} dataframe created by \code{simsigs()}, a function in a
forthcoming package \code{coherent}.}
\format{
A dataframe with the following fields:
\describe{
\item{Name}{Name of constellation used to create the record.}
\item{SNR}{The SNR in Decibles of the observation.}
\item{Bps}{The number of bits per symbol.  The number of bits in a
simulation run is \code{Bps * N}}
\item{NoisePower}{The actual noise power in the simulation run.  Since
the noise is randomly generated, this is a stochastic item.}
\item{N}{The number of symbols in the simulation run.}
\item{SER}{The number of symbols errors observed in the simulation run.}
\item{BER}{The number of bit errors observed in the simulation run.}
}
}
\usage{
BERDFc
}
\description{
\code{BERDF} is a standard \code{R} data frame created by the \code{simsigs()} function in
the forthcoming \code{coherent} package.  The observations have been \code{condense}d
}
\keyword{datasets}
