% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2BQ.R
\docType{package}
\name{binfunest-package}
\alias{binfunest}
\alias{binfunest-package}
\title{Package \code{B2BQ} will estimate an offset and B2B "Q" of a communications
system performance.}
\description{
Provides maximum likelihood estimates of the performance parameters that drive a binomial distribution of observed errors, and takes full advantage zero error observations. High performance communications systems typically have inherent noise sources and other performance limitations that need to be estimated. Measurements made at high signal to noise ratios typically result in zero errors due to limitation in available measurement time. Package includes theoretical performance functions for common modulation schemes, and utility functions to work with the performance functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PhilShea/binfunest}
  \item Report bugs at \url{https://github.com/PhilShea/binfunest/issues}
}

}
\author{
\strong{Maintainer}: Philip Shea \email{philshea@gmail.com} (\href{https://orcid.org/0000-0001-5460-9987}{ORCID})

}
\keyword{internal}
