% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor.mle}
\alias{cor.mle}
\title{cor.mle returns the correlation matrix of an mle fit.}
\usage{
cor.mle(m)
}
\arguments{
\item{m}{an \code{mle} object with a fit of more than one variable.}
}
\value{
a symmetric matrix with ones on the diagonal.
}
\description{
This simple function will return the correlation matrix for \code{mle} fits of
more than one variable.  Note that this function is \emph{not} an S3 generic.
}
\examples{
Q_ab <- function( gamma, a, b)
cor.mle( mle1)

}
\seealso{
\code{\link[stats4:mle]{stats4::mle()}}
}
