% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B2BConvert.R
\name{B2BConvert}
\alias{B2BConvert}
\title{B2BConvert Converts a function of SNR into one of SNR, B2B, and Offset.}
\usage{
B2BConvert(f)
}
\arguments{
\item{f}{A function of a single argument \code{f( s)}.}
}
\value{
A function of three arguments \code{f( s, B2B, offset)}..
}
\description{
Creates a function \code{f( -dB( undB( -s) + undB( -B2B)) - offset)}
}
\details{
Note that all quantities are assumed to be in Decibels.
}
\examples{
QPSKdB.B2B <- B2BConvert( QPSKdB)

}
